[navernow] Changes for feedback received
